chore(deps): batch 10 dependabot bumps into one PR
Consolidates the 10 open dependabot dependency-bump PRs (#339, #345-#353)
into a single change, with the source migrations each major bump requires.

Bumps:
- safetensors 0.7 -> 0.8
- petgraph 0.7 -> 0.8
- rustworkx-core 0.16 -> 0.17
- sha2 0.10 -> 0.11
- hmac 0.12 -> 0.13
- tree-sitter 0.24 -> 0.26
- openssl-sys 0.9.116 -> 0.9.117
- regex 1.12.3 -> 1.12.4
- cc 1.2.61 -> 1.2.64
- tauri 2.11.1 -> 2.11.2

Source migrations required by the major bumps:
- sha2 0.11 changed finalize()'s output type (no longer impls LowerHex):
  src/reasoning/cache.rs now uses hex::encode() — same lowercase hex output.
- hmac 0.13 moved new_from_slice() to the KeyInit trait:
  src/runner/providers/webhook.rs imports hmac::digest::KeyInit.
- tree-sitter 0.26 changed Node::child() to take u32 (was usize):
  src/parser/languages/dart.rs casts the index. The 0.23 grammar crates work
  unchanged with core 0.26 via the tree-sitter-language abstraction.

Cargo.lock note: re-resolution realigns async-nats (0.49.1 -> 0.47.0) and
bcrypt (0.19.1 -> 0.16.0) to match the existing Cargo.toml requirements
(async-nats = "0.47", bcrypt = "0.16"). The committed lock was stale/ahead of
Cargo.toml; the build already used the Cargo.toml versions, so this is a lock
correction, not a functional downgrade.

Validation: cargo check --lib, clippy --all-targets --all-features -D warnings,
cargo fmt --all --check, and the webhook HMAC signature tests all pass.

Closes #339, #345, #346, #347, #348, #349, #350, #351, #352, #353.

+++ src/parser/languages/dart.rs
@@ -44,7 +44,8 @@ fn extract_toplevel(
     let mut i = 0;
 
     while i < child_count {
-        let child = root.child(i).unwrap();
+        // tree-sitter 0.26 changed Node::child() to take u32 (was usize).
+        let child = root.child(i as u32).unwrap();
 
         match child.kind() {
             "class_declaration" => {
@@ -95,7 +96,8 @@ fn extract_toplevel(
             "function_signature" => {
                 // Look ahead for function_body
                 let body_node = if i + 1 < child_count {
-                    root.child(i + 1).filter(|n| n.kind() == "function_body")
+                    root.child((i + 1) as u32)
+                        .filter(|n| n.kind() == "function_body")
                 } else {
                     None
                 };
